+++ LearningPlan.Web.v2/src/components/Signup.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { useAuthentication, AuthenticationMethods, AuthenticationStatus } from "@/models/useAuthentication"
+import { useAuthentication, AuthenticationMethods, AuthenticationStatus } from "@/modules/authentication/authentication.controller";
 import { type Ref, ref, watch } from "vue";
 import LoadingWheelVue from "./icons/LoadingWheel.vue";
 import ErrorPopupVue from "./ErrorPopup.vue";
