+++ LearningPlan.Web.v2/src/modules/topic/classes/topic.class.ts
